+++ forge/src/main/java/com/binaris/wizardry/gametest/ForgeArcaneWorkbenchTest.java
@@ -100,6 +100,9 @@ public class ForgeArcaneWorkbenchTest {
     public static void cannotExceedUpgradeLimit(GameTestHelper helper) {
         List<Item> upgrades = List.of(
                 EBItems.ARCANE_TOME.get(),
+                EBItems.APPRENTICE_ARCANE_TOME.get(),
+                EBItems.ADVANCED_ARCANE_TOME.get(),
+                EBItems.MASTER_ARCANE_TOME.get(),
                 EBItems.CRYSTAL_SILVER_PLATING.get(),
                 EBItems.ETHEREAL_CRYSTAL_WEAVE.get(),
                 EBItems.RESPLENDENT_THREAD.get()
@@ -107,6 +110,13 @@ public class ForgeArcaneWorkbenchTest {
 
         upgrades.forEach(upgrade ->
                 ArcaneWorkbenchTest.cannotExceedUpgradeLimit(helper, upgrade));
+        helper.succeed();
+    }
+
+    @GameTest(template = "arcane_workbench_3x3x3")
+    public static void upgradeWandNextTierNBT(GameTestHelper helper) {
+        EBLogger.warn("upgradeWandNextTierNBT not implemented on Forge due to ServerPlayer cast limitations.");
+        //        EBDataGenProcessor.wandItems().values().forEach(wand -> ArcaneWorkbenchTest.upgradeWandNextTierNBT(helper, wand.get()));
         helper.succeed();
     }
 }