+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Fixed ring of the Desert Sun wasn't marked as a ring, so you couldn't use it in the ring slot.
 - Fixed that minions only had half of their intended lifetime, so they would die much earlier than intended.
+- Added arcane tomes as individual items and not NBT dependant. Now giving the option to use NBT or just regular arcane tomes with new tiers added by addons.